+++ src/main/java/ml/pluto7073/bartending/foundations/datagen/BartendingDrinks.java
@@ -1,6 +1,5 @@
 package ml.pluto7073.bartending.foundations.datagen;
 
-import ml.pluto7073.bartending.TheArtOfBartending;
 import ml.pluto7073.bartending.content.item.BartendingItems;
 import ml.pluto7073.pdapi.PDAPI;
 import ml.pluto7073.pdapi.addition.action.ApplyStatusEffectAction;
@@ -8,11 +7,8 @@ import ml.pluto7073.pdapi.addition.action.RestoreHungerAction;
 import ml.pluto7073.pdapi.datagen.provider.SpecialtyDrinkProvider;
 import ml.pluto7073.pdapi.specialty.SpecialtyDrink;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
-import net.fabricmc.fabric.api.resource.conditions.v1.ConditionJsonProvider;
 import net.fabricmc.fabric.api.resource.conditions.v1.DefaultResourceConditions;
-import net.fabricmc.fabric.impl.datagen.FabricDataGenHelper;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffects;
 
 import java.util.function.BiConsumer;
@@ -50,7 +46,7 @@ public class BartendingDrinks extends SpecialtyDrinkProvider {
             .step(asId("shot_of_tequila"))
             .step(asId("shot_of_orange_liqueur"))
             .step(asId("compat/fruitfulfun/lime"))
-            .chemical(asId("alcohol"), 34)
+            .chemical(asId("alcohol"), 42)
             .color(14808274)
             .action(new ApplyStatusEffectAction(MobEffects.FIRE_RESISTANCE, 200, 1))
             .action(new RestoreHungerAction(3, 3));
@@ -67,6 +63,50 @@ public class BartendingDrinks extends SpecialtyDrinkProvider {
             .color(2167556)
             .action(new RestoreHungerAction(3, 0));
 
+    private static final DrinkBuilder DEATH_IN_THE_AFTERNOON = staticBaseBuilder(BartendingItems.GLASS_OF_CHAMPAGNE)
+            .step(asId("shot_of_absinthe"))
+            .chemical(asId("alcohol"), 34)
+            .color(0xe0de50);
+
+    private static final DrinkBuilder OLD_FASHIONED = staticBaseBuilder(BartendingItems.MIXED_DRINK)
+            .step(PDAPI.asId("sugar"))
+            .step(asId("shot_of_whiskey"))
+            .step(asId("sweet_berries"))
+            .chemical(asId("alcohol"), 14)
+            .color(0x442612)
+            .action(new RestoreHungerAction(1, 1));
+
+    private static final DrinkBuilder MIMOSA = staticBaseBuilder(BartendingItems.GLASS_OF_CHAMPAGNE)
+            .step(asId("compat/fruitfulfun/orange"))
+            .chemical(asId("alcohol"), 14)
+            .color(0xf2c041)
+            .action(new RestoreHungerAction(3, 1));
+
+    private static final DrinkBuilder KAMIKAZE = staticBaseBuilder(BartendingItems.MIXED_DRINK)
+            .step(asId("shot_of_vodka"))
+            .step(asId("shot_of_orange_liqueur"))
+            .step(asId("compat/fruitfulfun/lime"))
+            .chemical(asId("alcohol"), 28)
+            .color(0xe0f9db)
+            .action(new RestoreHungerAction(3, 3));
+
+    private static final DrinkBuilder MANHATTAN = staticBaseBuilder(BartendingItems.MIXED_DRINK)
+            .step(asId("shot_of_whiskey"))
+            .step(asId("shot_of_whiskey"))
+            .step(asId("shot_of_sweet_vermouth"))
+            .chemical(asId("alcohol"), 34f)
+            .color(0xc46836);
+
+    private static final DrinkBuilder VODKA_MARTINI = staticBaseBuilder(BartendingItems.MIXED_DRINK)
+            .step(PDAPI.asId("ice"))
+            .step(asId("shot_of_vodka"))
+            .step(asId("shot_of_vodka"))
+            .step(asId("shot_of_vodka"))
+            .step(asId("shot_of_dry_vermouth"))
+            .chemical(asId("alcohol"), 48)
+            .color(15724527)
+            .action(new ApplyStatusEffectAction(MobEffects.FIRE_RESISTANCE, 600, 1));
+
     public BartendingDrinks(FabricDataOutput out) {
         super(out);
     }
@@ -79,6 +119,12 @@ public class BartendingDrinks extends SpecialtyDrinkProvider {
         MARTINI.save(asId("martini"), output);
         MARGARITA.save(asId("compat/fruitfulfun/margarita"), fruitfulfun);
         ESPRESSO_MARTINI.save(asId("compat/plutoscoffee/espresso_martini"), plutoscoffee);
+        DEATH_IN_THE_AFTERNOON.save(asId("death_in_the_afternoon"), output);
+        OLD_FASHIONED.save(asId("old_fashioned"), output);
+        MIMOSA.save(asId("compat/fruitfulfun/mimosa"), fruitfulfun);
+        KAMIKAZE.save(asId("compat/fruitfulfun/kamikaze"), fruitfulfun);
+        MANHATTAN.save(asId("manhattan"), output);
+        VODKA_MARTINI.save(asId("vodka_martini"), output);
     }
 
 }